Split offscreen fallback media overlays
